% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_news_detection.R
\name{fake_news_detection}
\alias{fake_news_detection}
\title{Fake News Detection: Clickbait & Similarity Analysis}
\usage{
fake_news_detection(news)
}
\arguments{
\item{news}{A dataframe containing output from get_news_everything() method}
}
\value{
A dataframe with the chosen analysis results
}
\description{
This function allows users to choose between clickbait detection and content similarity analysis.
Users can also choose between using OpenAI API or R-based NLP analysis.
}
\examples{
\dontrun{

# Get articles
news <- get_news_everything(
 keyword = "Apple",
 from = "2025-03-01",
 to = "2025-03-11",
 sortBy = "popularity"
)

# Run function
result <- fake_news_detection(news)
print(result)

}
}
