% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_article.R
\name{get_full_article}
\alias{get_full_article}
\title{Get the Full Article Text from a URL and Remove the Last Sentence}
\usage{
get_full_article(url)
}
\arguments{
\item{url}{A character string containing the URL of the article.}
}
\value{
A character string containing the cleaned article text with the last sentence removed.
}
\description{
This function retrieves the full text of an article from the given URL by extracting
the content of \verb{<p>} tags. It then performs several cleanup steps.
Finally, it removes the last two sentence of the article, under the assumption that it is
generally not important.
}
