% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_translation.R
\name{get_openai_response}
\alias{get_openai_response}
\title{Get OpenAI API Response}
\usage{
get_openai_response(
  messages = NULL,
  user_msg = NULL,
  system_msg =
    "You are a professional news assistant. Provide exactly what the user requests without any extra commentary.",
  model = "gpt-3.5-turbo",
  temperature = 0.2
)
}
\arguments{
\item{messages}{A character string containing a customed conversation.}

\item{user_msg}{A character string containing the user's prompt.}

\item{system_msg}{A character string containing the system's prompt.}

\item{model}{A character string specifying the model to use (default: "gpt-3.5-turbo").}

\item{temperature}{A float meaning randomness.}
}
\value{
A character string with the content of the response message.
}
\description{
This function sends a request to the OpenAI Chat Completions endpoint
using the specified prompt and parameters. It returns the response message content.
}
