% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_api.R
\name{get_top_headlines}
\alias{get_top_headlines}
\title{Get live top headlines from NewsAPI "top-headlines" Endpoint}
\usage{
get_top_headlines(category = NULL)
}
\arguments{
\item{category}{A character string representing the category you want to get headlines for. (e.g., :business",
"entertainment", "general", "health", "science", "sports", "technology")
If NULL, this parameter is omitted from the query.}
}
\value{
A data frame containing the top headlines.
}
\description{
Get live top headlines from NewsAPI "top-headlines" Endpoint
}
\examples{
\dontrun{
  # Retrieve headlines from all categories
  headlines <- get_top_headlines()

  # Retrieve headlines from the "technology" category
  tech_headlines <- get_top_headlines(category = "technology")
}

}
