% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_analysis.R
\name{plot_word_cloud}
\alias{plot_word_cloud}
\title{Plot Word Cloud}
\usage{
plot_word_cloud(articles, text_column = "content", max_words = 100)
}
\arguments{
\item{articles}{A data frame containing the articles.}

\item{text_column}{A character string indicating the column name that contains the text (default: "content").}

\item{max_words}{An integer specifying the maximum number of words to display in the word cloud (default: 100).}
}
\value{
None. The function prints the word cloud.
}
\description{
This function generates a word cloud from the text column in the articles data frame.
It tokenizes the text, removes stopwords, calculates word frequencies, and plots the word cloud.
}
\examples{
\dontrun{
plot_word_cloud(analyzed_articles, text_column = "content", max_words = 50)
}

}
